Refactor, and actually filter for Trace
